Demo.Mmose.Core.Test/ServerEngine/World/Domain/DomainTest.cs
using Demo.Mmose.Core.World.DomainWorld;
using Microsoft.VisualStudio.TestTools.UnitTesting;
using Demo.Mmose.Core.World;
using Demo.Mmose.Core.Network;

namespace Demo.Mmose.Core.Test
{
    /// <summary>
    ///这是 DomainTest 的测试类，旨在
    ///包含所有 DomainTest 单元测试
    ///</summary>
    [TestClass()]
    public class DomainTest
    {
        private TestContext testContextInstance;
        /// <summary>
        ///获取或设置测试上下文，上下文提供
        ///有关当前测试运行及其功能的信息。
        ///</summary>
        public TestContext TestContext
        {
            get { return testContextInstance; }
            set { testContextInstance = value; }
        }

        #region 附加测试属性
        // 
        //编写测试时，还可使用以下属性:
        //
        //使用 ClassInitialize 在运行类中的第一个测试前先运行代码
        //[ClassInitialize()]
        //public static void MyClassInitialize(TestContext testContext)
        //{
        //}
        //
        //使用 ClassCleanup 在运行完类中的所有测试后再运行代码
        //[ClassCleanup()]
        //public static void MyClassCleanup()
        //{
        //}
        //
        //使用 TestInitialize 在运行每个测试前先运行代码
        //[TestInitialize()]
        //public void MyTestInitialize()
        //{
        //}
        //
        //使用 TestCleanup 在运行完每个测试后运行代码
        //[TestCleanup()]
        //public void MyTestCleanup()
        //{
        //}
        //
        #endregion

        /// <summary>
        ///ZoneClusterHandler 的测试
        ///</summary>
        public void ZoneClusterHandlerTestHelper<T>()
            where T : BaseWorld, new()
        {
            Domain<T> target = new Domain<T>(); // TODO: 初始化为适当的值
            ZoneClusterHandler actual;
            actual = target.ZoneClusterHandler;
            Assert.Inconclusive( "验证此测试方法的正确性。" );
        }

        [TestMethod()]
        public void ZoneClusterHandlerTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 ZoneClusterHandlerTestHelper<T>()。" );
        }

        /// <summary>
        ///World 的测试
        ///</summary>
        public void WorldTestHelper<T>()
            where T : BaseWorld, new()
        {
            Domain<T> target = new Domain<T>(); // TODO: 初始化为适当的值
            T actual;
            actual = target.World;
            Assert.Inconclusive( "验证此测试方法的正确性。" );
        }

        [TestMethod()]
        public void WorldTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 WorldTestHelper<T>()。" );
        }

        /// <summary>
        ///ProcessorBuffers 的测试
        ///</summary>
        public void ProcessorBuffersTestHelper<T>()
            where T : BaseWorld, new()
        {
            Domain_Accessor<T> target = new Domain_Accessor<T>(); // TODO: 初始化为适当的值
            BufferPool actual;
            actual = target.ProcessorBuffers;
            Assert.Inconclusive( "验证此测试方法的正确性。" );
        }

        [TestMethod()]
        [DeploymentItem( "Demo.Mmose.Core.dll" )]
        public void ProcessorBuffersTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 ProcessorBuffersTestHelper<T>()。" );
        }

        /// <summary>
        ///IsLoginDomain 的测试
        ///</summary>
        public void IsLoginDomainTestHelper<T>()
            where T : BaseWorld, new()
        {
            Domain<T> target = new Domain<T>(); // TODO: 初始化为适当的值
            bool actual;
            actual = target.IsLoginDomain;
            Assert.Inconclusive( "验证此测试方法的正确性。" );
        }

        [TestMethod()]
        public void IsLoginDomainTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 IsLoginDomainTestHelper<T>()。" );
        }

        /// <summary>
        ///DomainPacketHandlers 的测试
        ///</summary>
        public void DomainPacketHandlersTestHelper<T>()
            where T : BaseWorld, new()
        {
            Domain<T> target = new Domain<T>(); // TODO: 初始化为适当的值
            DomainPacketHandlers actual;
            actual = target.DomainPacketHandlers;
            Assert.Inconclusive( "验证此测试方法的正确性。" );
        }

        [TestMethod()]
        public void DomainPacketHandlersTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 DomainPacketHandlersTestHelper<T>()。" );
        }

        /// <summary>
        ///ConfigDomain 的测试
        ///</summary>
        public void ConfigDomainTestHelper<T>()
            where T : BaseWorld, new()
        {
            Domain<T> target = new Domain<T>(); // TODO: 初始化为适当的值
            ConfigDomain actual;
            actual = target.ConfigDomain;
            Assert.Inconclusive( "验证此测试方法的正确性。" );
        }

        [TestMethod()]
        public void ConfigDomainTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 ConfigDomainTestHelper<T>()。" );
        }

        /// <summary>
        ///ListenerProcessReceive 的测试
        ///</summary>
        public void ListenerProcessReceiveTestHelper<T>()
            where T : BaseWorld, new()
        {
            Domain<T> target = new Domain<T>(); // TODO: 初始化为适当的值
            object sender = null; // TODO: 初始化为适当的值
            NetStateEventArgs eventArgs = null; // TODO: 初始化为适当的值
            target.ListenerProcessReceive( sender, eventArgs );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        public void ListenerProcessReceiveTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 ListenerProcessReceiveTestHelper<T>()。" );
        }

        /// <summary>
        ///ListenerInitNetState 的测试
        ///</summary>
        public void ListenerInitNetStateTestHelper<T>()
            where T : BaseWorld, new()
        {
            Domain<T> target = new Domain<T>(); // TODO: 初始化为适当的值
            object sender = null; // TODO: 初始化为适当的值
            NetStateInitEventArgs netStateInit = null; // TODO: 初始化为适当的值
            target.ListenerInitNetState( sender, netStateInit );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        public void ListenerInitNetStateTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 ListenerInitNetStateTestHelper<T>()。" );
        }

        /// <summary>
        ///ListenerDomainServer 的测试
        ///</summary>
        public void ListenerDomainServerTestHelper<T>()
            where T : BaseWorld, new()
        {
            Domain_Accessor<T> target = new Domain_Accessor<T>(); // TODO: 初始化为适当的值
            bool expected = false; // TODO: 初始化为适当的值
            bool actual;
            actual = target.ListenerDomainServer();
            Assert.AreEqual( expected, actual );
            Assert.Inconclusive( "验证此测试方法的正确性。" );
        }

        [TestMethod()]
        [DeploymentItem( "Demo.Mmose.Core.dll" )]
        public void ListenerDomainServerTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 ListenerDomainServerTestHelper<T>()。" );
        }

        /// <summary>
        ///InitDomain 的测试
        ///</summary>
        public void InitDomainTestHelper<T>()
            where T : BaseWorld, new()
        {
            Domain<T> target = new Domain<T>(); // TODO: 初始化为适当的值
            ConfigDomain configDomain = null; // TODO: 初始化为适当的值
            target.InitDomain( configDomain );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        public void InitDomainTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 InitDomainTestHelper<T>()。" );
        }

        /// <summary>
        ///GetPacketLength 的测试
        ///</summary>
        public void GetPacketLengthTestHelper<T>()
            where T : BaseWorld, new()
        {
            Domain_Accessor<T> target = new Domain_Accessor<T>(); // TODO: 初始化为适当的值
            object sender = null; // TODO: 初始化为适当的值
            PacketLengthInfoEventArgs eventArgs = null; // TODO: 初始化为适当的值
            target.GetPacketLength( sender, eventArgs );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        [DeploymentItem( "Demo.Mmose.Core.dll" )]
        public void GetPacketLengthTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 GetPacketLengthTestHelper<T>()。" );
        }

        /// <summary>
        ///GetPacketID 的测试
        ///</summary>
        public void GetPacketIDTestHelper<T>()
            where T : BaseWorld, new()
        {
            Domain_Accessor<T> target = new Domain_Accessor<T>(); // TODO: 初始化为适当的值
            object sender = null; // TODO: 初始化为适当的值
            PacketIdInfoEventArgs eventArgs = null; // TODO: 初始化为适当的值
            target.GetPacketID( sender, eventArgs );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        [DeploymentItem( "Demo.Mmose.Core.dll" )]
        public void GetPacketIDTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 GetPacketIDTestHelper<T>()。" );
        }

        /// <summary>
        ///DomainWorld_InitOnceServer 的测试
        ///</summary>
        public void DomainWorld_InitOnceServerTestHelper<T>()
            where T : BaseWorld, new()
        {
            Domain_Accessor<T> target = new Domain_Accessor<T>(); // TODO: 初始化为适当的值
            object sender = null; // TODO: 初始化为适当的值
            BaseWorldEventArgs eventArgs = null; // TODO: 初始化为适当的值
            target.DomainWorld_InitOnceServer( sender, eventArgs );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        [DeploymentItem( "Demo.Mmose.Core.dll" )]
        public void DomainWorld_InitOnceServerTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 DomainWorld_InitOnceServerTestHelper<T>()。" );
        }

        /// <summary>
        ///DomainWorld_ExitWorld 的测试
        ///</summary>
        public void DomainWorld_ExitWorldTestHelper<T>()
            where T : BaseWorld, new()
        {
            Domain_Accessor<T> target = new Domain_Accessor<T>(); // TODO: 初始化为适当的值
            object sender = null; // TODO: 初始化为适当的值
            BaseWorldEventArgs eventArgs = null; // TODO: 初始化为适当的值
            target.DomainWorld_ExitWorld( sender, eventArgs );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        [DeploymentItem( "Demo.Mmose.Core.dll" )]
        public void DomainWorld_ExitWorldTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 DomainWorld_ExitWorldTestHelper<T>()。" );
        }

        /// <summary>
        ///ConnecterProcessReceive 的测试
        ///</summary>
        public void ConnecterProcessReceiveTestHelper<T>()
            where T : BaseWorld, new()
        {
            Domain<T> target = new Domain<T>(); // TODO: 初始化为适当的值
            object sender = null; // TODO: 初始化为适当的值
            NetStateEventArgs eventArgs = null; // TODO: 初始化为适当的值
            target.ConnecterProcessReceive( sender, eventArgs );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        public void ConnecterProcessReceiveTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 ConnecterProcessReceiveTestHelper<T>()。" );
        }

        /// <summary>
        ///ConnecterIniNetState 的测试
        ///</summary>
        public void ConnecterIniNetStateTestHelper<T>()
            where T : BaseWorld, new()
        {
            Domain<T> target = new Domain<T>(); // TODO: 初始化为适当的值
            object sender = null; // TODO: 初始化为适当的值
            NetStateInitEventArgs netStateInit = null; // TODO: 初始化为适当的值
            target.ConnecterIniNetState( sender, netStateInit );
            Assert.Inconclusive( "无法验证不返回值的方法。" );
        }

        [TestMethod()]
        public void ConnecterIniNetStateTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 ConnecterIniNetStateTestHelper<T>()。" );
        }

        /// <summary>
        ///Domain`1 构造函数 的测试
        ///</summary>
        public void DomainConstructorTestHelper<T>()
            where T : BaseWorld, new()
        {
            Domain<T> target = new Domain<T>();
            Assert.Inconclusive( "TODO: 实现用来验证目标的代码" );
        }

        [TestMethod()]
        public void DomainConstructorTest()
        {
            Assert.Inconclusive( "没有找到能够满足 T 的类型约束的相应类型参数。请以适当的类型参数来调用 DomainConstructorTestHelper<T>()。" );
        }
    }
}
